fix(build-tool,ci): close two blind spots that hid clippy errors, and fix what they hid
The clippy gate has two holes that let `-D warnings` errors sit on main
undetected. Both are fixed here, along with every error they were hiding.

Hole 1 — a BUILD preamble switched the gate off for that package.
`executor.clippyStepFor` inspected only `buildCommands[0]` when deciding
whether a Rust package runs cargo. `readLines` strips `#!/bin/sh` and
`#`-comments but not `set -e`, `export VAR=…`, `cd …` or progress `echo`s,
so a BUILD like sql-codegen's presented `set -e` as command zero, matched
no cargo, and got no clippy step at all — on any platform, on any leg. 22
packages were dark this way. `coding-adventures-sql-planner` had a live
`manual_is_multiple_of` error on main because of it.

The derivation now scans every command and prefers an unconditional cargo
invocation over a platform-guarded one. Scanning all lines opens the
mirror-image trap — compile-only crates *print* the command they cannot
run (`echo "  To build: cargo build -p font-parser-node --release"`) — so
`runsCargo` erases single- and double-quoted spans before looking for a
bare `cargo` word. `RUSTDOCFLAGS="-D warnings" cargo doc` and
`cd "$WORKSPACE" && cargo test` still count; the echo does not.

Hole 2 — a crate that `compile_error!`s off Windows is linted by nobody.
Its default BUILD is a bare `echo SKIP`, so the Linux and macOS legs
correctly emit no clippy step; its real commands live in BUILD_windows,
which never runs because CI skips the build step on the Windows leg. That
is how `paint-vm-direct2d` accumulated 8 errors (PR #12119) — and because
clippy lints path dependencies, those 8 broke `cargo clippy --all-targets
-- -D warnings` for every crate depending on it while its own gate step
reported success. ci.yml gains a `Clippy Windows-only Rust crates` step on
the Windows PR leg. The crate list is derived by grepping for the
`#[cfg(not(target_os = "windows"))] compile_error!` declaration itself, so
a new Windows-only crate joins the gate automatically, and the step fails
loudly if the derived set is ever empty.

Errors the two fixes exposed, all fixed here:

- paint-vm-gdi (3): `unnecessary_cast` ×2 on `BI_RGB.0 as u32` — both
  sides are already `u32`; `too_many_arguments` ×1 on
  `render_offscreen_composited`, silenced with a scoped allow and a
  documented reason (its two callers read the middle parameter group off
  `PaintGroup` and `PaintLayer`, so there is no honest struct to bundle).
  Audited for the `.min(255)`-after-`as u8` idiom that produced three of
  paint-vm-direct2d's errors: it does not occur, and cannot — this backend
  never un-premultiplies, so there is no quotient that can exceed 255.
- sql-planner (1): `manual_is_multiple_of`.
- sql-vm (2): `approx_constant` on two test literals, resolved with scoped
  allows per the repo's standing rule for that lint.
- polynomial-wasm (11): `missing_safety_doc` on every `pub unsafe extern
  "C"` export. Each now documents the provenance, length and
  initialization requirements on its `(ptr, len)` pairs.

Verification: the new CI step, run verbatim on Windows, reproduces
paint-vm-direct2d's 8 historical errors and catches a deliberately planted
`unnecessary_cast` in paint-vm-gdi (exit 101). `go vet ./...` and the full
build-tool suite pass; `cargo test --no-fail-fast` is green for
paint-vm-gdi, sql-planner and sql-vm.

+++ code/packages/rust/paint-vm-gdi/src/lib.rs
@@ -648,7 +648,13 @@ unsafe fn create_surface(width: u32, height: u32) -> Option<GdiSurface> {
             biHeight: -(height as i32),
             biPlanes: 1,
             biBitCount: 32,
-            biCompression: BI_RGB.0 as u32,
+            // `BI_RGB` is a `BI_COMPRESSION` newtype around a `u32`, and
+            // `biCompression` is itself a `u32` — so `.0` already has the
+            // right type and no cast is needed. (A `foo as u32` on a value
+            // that is already `u32` is what clippy's `unnecessary_cast`
+            // flags; it is also a small trap, because it would silently
+            // keep compiling if the underlying type ever changed width.)
+            biCompression: BI_RGB.0,
             biSizeImage: 0,
             biXPelsPerMeter: 0,
             biYPelsPerMeter: 0,
@@ -1022,7 +1028,32 @@ unsafe fn finalize_surface_alpha(color_surface: &mut GdiSurface, coverage_surfac
     }
 }
 
+/// Render `instructions` into an off-screen surface pair and alpha-blend the
+/// result onto `dest_hdc`.
+///
+/// ## Why eight parameters
+///
+/// Clippy's `too_many_arguments` (limit: 7) fires here, and the allow below is
+/// deliberate rather than lazy. The parameters split into three groups that do
+/// *not* travel together:
+///
+/// | Group                                      | Comes from                    |
+/// |--------------------------------------------|-------------------------------|
+/// | `dest_hdc`                                  | the caller's target surface   |
+/// | `instructions`, `transform`, `opacity`      | the group/layer being drawn   |
+/// | `scene_width`, `scene_height`, `gradients`, `mode` | the scene-wide render state |
+///
+/// Bundling the middle group into a struct would not work: the two call sites
+/// are `render_group` (which reads them off a `PaintGroup`) and `render_layer`
+/// (which reads them off a `PaintLayer`), and those are different types with
+/// different field names. Bundling only the scene-wide group into a context
+/// struct would ripple through `render_instructions` and every recursive
+/// renderer in this file — a large refactor of a Windows-only crate that CI
+/// cannot exercise, for no behavioural gain. So the argument list stays flat
+/// and the lint is silenced *here*, at the one function that needs it, with
+/// the reasoning recorded.
 #[cfg(target_os = "windows")]
+#[allow(clippy::too_many_arguments)]
 unsafe fn render_offscreen_composited(
     dest_hdc: HDC,
     instructions: &[PaintInstruction],
@@ -2018,7 +2049,9 @@ unsafe fn render_unsafe(scene: &PaintScene, width: u32, height: u32) -> PixelCon
             biHeight: -(height as i32), // Negative = top-down DIB
             biPlanes: 1,
             biBitCount: 32,
-            biCompression: BI_RGB.0 as u32,
+            // No cast: `BI_RGB.0` and `biCompression` are both `u32`.
+            // See the matching field in `create_surface` above.
+            biCompression: BI_RGB.0,
             biSizeImage: 0,
             biXPelsPerMeter: 0,
             biYPelsPerMeter: 0,
